requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.17.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0
spacy>=3.7.0
nltk>=3.8.0
